+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 
 # Carbon.IIS Changelog
 
+## 1.4.0
+
+### Added
+
+* The `Start-CIisAppPool` function now starts the Windows Process Activation Service (WAS) if it is not running.
+Application pools fail to start if WAS isn't running.
+* The `Start-CIisWebsite` function now starts the World Wide Web Publishing Service (W3SVC) if it is not running.
+  Websites fail to start if W3SVC isn't running.
+
+
 ## 1.3.0
 
 > Released 14 Nov 2023